Basic REU emulation support
This emulates a REU of fixed 16MiB size and reads the initial contents
from a file named 'reufile.linux' in the current directory.

FIXME: Proper cmdline arguments for this.
FIXME: There is seemingly a way to set up the REU to keep the register
contents (Which is also a TODO for semu-c64) and/or do autoincrements.
This is all not yet implemented, the REU only emulates what semu-c64
needs at the moment.