Make BASIN() peek for char, as that is the same behavior running on Vice
BASIN() now returns zero when nothing is available. This seems to be the
behavior on Vice when using the llvm-mos runtime on top, but not necessarily
the true behavior of a real C64!

Note that there is input conversion still missing, you meed need to type
uppercase for lowercase and vice versa. Also, proper (non-)ECHO handling
doesn't work yet (FIXME). Also, the terminal might need resetting after
running the emulator with atexit(..) or similar, didn't check that yet
(FIXME#2).